requirement.txt
beautifulsoup4
requests
pillow
psycopg2
fuzzywuzzy